more cleanup, working on console errors
